+++ UnityEngine.TestRunner/Assertions/LogAssert.cs
@@ -28,6 +28,10 @@ namespace UnityEngine.TestTools
             }
             set
             {
+                if (value != LogScope.Current.IgnoreFailingMessages)
+                {
+                    Debug.LogFormat(LogType.Log, LogOption.NoStacktrace, null, "\nIgnoreFailingMessages:" + (value? "true":"false"));
+                }
                 LogScope.Current.IgnoreFailingMessages = value;
             }
         }