+++ UnityEngine.TestRunner/NUnitExtensions/Attributes/UnityTestAttribute.cs
@@ -9,7 +9,7 @@ using UnityEngine.TestRunner.NUnitExtensions.Runner;
 namespace UnityEngine.TestTools
 {
     [AttributeUsage(AttributeTargets.Method)]
-    public class UnityTestAttribute : CombiningStrategyAttribute, IWrapSetUpTearDown, ISimpleTestBuilder, IImplyFixture
+    public class UnityTestAttribute : CombiningStrategyAttribute, ISimpleTestBuilder, IImplyFixture
     {
         public UnityTestAttribute() : base(new UnityCombinatorialStrategy(), new ParameterDataSourceProvider()) {}
 
@@ -28,16 +28,6 @@ namespace UnityEngine.TestTools
             if (t.parms != null)
                 t.parms.HasExpectedResult = false;
             return t;
-        }
-
-        public TestCommand Wrap(TestCommand command)
-        {
-            return new OuterUnityTestActionCommand(
-                new EnumerableSetUpTearDownCommand(
-                    new SetUpTearDownCommand(
-                        new TestActionCommand(
-                            new UnityLogCheckDelegatingCommand(
-                                new EnumerableTestMethodCommand((TestMethod)command.Test))))));
         }
     }
 }