+++ UnityEngine.TestRunner/NUnitExtensions/Runner/DefaultTestWorkItem.cs
@@ -23,48 +23,7 @@ namespace UnityEngine.TestRunner.NUnitExtensions.Runner
         public DefaultTestWorkItem(TestMethod test, ITestFilter filter)
             : base(test, null)
         {
-            _command = test.RunState == RunState.Runnable || test.RunState == RunState.Explicit && filter.IsExplicitMatch(test)
-                ? BuildTestCommand(test)
-                : new SkipCommand(test);
-        }
-
-        private static TestCommand BuildTestCommand(TestMethod test)
-        {
-            var command = (TestCommand)new TestMethodCommand(test);
-            command = new UnityLogCheckDelegatingCommand(command);
-            foreach (var wrapper in test.Method.GetCustomAttributes<IWrapTestMethod>(true))
-            {
-                command = wrapper.Wrap(command);
-                if (command == null)
-                {
-                    var message = string.Format("IWrapTestMethod implementation '{0}' returned null as command.", wrapper.GetType().FullName);
-                    return new FailCommand(test, ResultState.Failure, message);
-                }
-            }
-
-            command = new TestTools.TestActionCommand(command);
-            command = new TestTools.SetUpTearDownCommand(command);
-            command = new ImmediateEnumerableCommand(command);
-            foreach (var wrapper in test.Method.GetCustomAttributes<IWrapSetUpTearDown>(true))
-            {
-                command = wrapper.Wrap(command);
-                if (command == null)
-                {
-                    var message = string.Format("IWrapSetUpTearDown implementation '{0}' returned null as command.", wrapper.GetType().FullName);
-                    return new FailCommand(test, ResultState.Failure, message);
-                }
-            }
-
-            command = new EnumerableSetUpTearDownCommand(command);
-            command = new OuterUnityTestActionCommand(command);
-
-            IApplyToContext[] changes = test.Method.GetCustomAttributes<IApplyToContext>(true);
-            if (changes.Length > 0)
-            {
-                command = new EnumerableApplyChangesToContextCommand(command, changes);
-            }
-
-            return command;
+            _command = TestCommandBuilder.BuildTestCommand(test, filter);
         }
 
         protected override IEnumerable PerformWork()