+++ UnityEngine.TestRunner/NUnitExtensions/Runner/FailCommand.cs
@@ -1,11 +1,12 @@
 
+using System.Collections;
 using NUnit.Framework.Interfaces;
 using NUnit.Framework.Internal;
 using NUnit.Framework.Internal.Commands;
 
 namespace UnityEngine.TestRunner.NUnitExtensions.Runner
 {
-    internal class FailCommand : TestCommand
+    internal class FailCommand : TestCommand, IEnumerableTestMethodCommand
     {
         private ResultState m_ResultState;
         private string m_Message;
@@ -21,6 +22,12 @@ namespace UnityEngine.TestRunner.NUnitExtensions.Runner
         {
             context.CurrentResult.SetResult(m_ResultState, m_Message);
             return context.CurrentResult;
+        }
+
+        public IEnumerable ExecuteEnumerable(ITestExecutionContext context)
+        {
+            context.CurrentResult.SetResult(m_ResultState, m_Message);
+            yield return null;
         }
     }